pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "leadgen"
version = "0.1.0"
description = "AI-powered lead generation engine — aka Legion"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    # AI / LLM
    "anthropic>=0.30.0",

    # HTTP & APIs
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # Web crawling
    "playwright>=1.44.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",

    # Data handling
    "pydantic>=2.7.0",
    "pandas>=2.2.0",

    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",

    # Email
    "aiosmtplib>=3.0.0",
    "sendgrid>=6.11.0",

    # MCP server
    "mcp>=1.0.0",

    # Config & env
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",

    # Utilities
    "rich>=13.7.0",
    "tenacity>=8.3.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
leadgen = "leadgen.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/leadgen"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
